fpga: dfl: move dfl-bus related APIs to include/linux/fpga/dfl-bus.h
The patch moves dfl-bus related APIs to include/linux/fpga/dfl-bus.h

Now the DFL sub feature drivers could be made as independent modules and
put in different folders according to their functionality. In order for
scattered sub feature drivers to include dfl bus APIs, move the dfl bus
APIs to a new header file in the public folder.
